Add redux and conditional rendering middleware
Collections are stored in the global state (models to come later). This allows
for rendering when the collections are changed. Also added is a middleware
that allows for conditional rendering by registering collections and models
in the state and allowing the view to check to see if any actions relating
to it were dispatched.